pyproject.toml
[project]
name = "ai-tools"
version = "0.1.0"
description = "Personal AI tools: chat UI, audio/video utilities, news digest, and analysis scripts."
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "aiofiles==23.2.1",
    "aiohappyeyeballs==2.4.6",
    "aiohttp==3.11.12",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "arxiv==2.1.3",
    "async-timeout==5.0.1",
    "attrs==25.1.0",
    "azure-core==1.32.0",
    "azure-identity==1.20.0",
    "bcrypt==4.2.1",
    "beautifulsoup4==4.13.3",
    "cachetools==5.5.1",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "composio_core==0.7.1",
    "composio_gemini==0.7.1.post1",
    "composio_llamaindex==0.7.1",
    "cryptography==44.0.0",
    "dataclasses-json==0.6.7",
    "Deprecated==1.2.18",
    "dirtyjson==1.0.8",
    "distro==1.9.0",
    "exceptiongroup==1.2.2",
    "fastapi==0.115.8",
    "feedparser==6.0.11",
    "ffmpy==0.5.0",
    "filelock==3.17.0",
    "filetype==1.2.0",
    "frozenlist==1.5.0",
    "fsspec==2025.2.0",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.24.1",
    "google-api-python-client==2.160.0",
    "google-auth==2.38.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.1",
    "google-genai==1.0.0",
    "google-generativeai==0.8.4",
    "googleapis-common-protos==1.67.0rc1",
    "gradio==5.16.1",
    "gradio_client==1.7.0",
    "greenlet==3.1.1",
    "grpcio==1.70.0",
    "grpcio-status==1.70.0",
    "h11==0.14.0",
    "html2text==2024.2.26",
    "httpcore==1.0.7",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "huggingface-hub==0.28.1",
    "idna==3.10",
    "importlib_metadata==8.6.1",
    "inflection==0.5.1",
    "Jinja2==3.1.5",
    "jiter==0.8.2",
    "joblib==1.4.2",
    "jsonref==1.1.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "llama-cloud==0.1.12",
    "llama-cloud-services==0.6.0",
    "llama-hub==0.0.79.post1",
    "llama-index==0.12.16",
    "llama-index-agent-openai==0.4.5",
    "llama-index-cli==0.4.0",
    "llama-index-core==0.12.16.post1",
    "llama-index-embeddings-openai==0.3.1",
    "llama-index-indices-managed-llama-cloud==0.6.4",
    "llama-index-llms-azure-openai==0.3.0",
    "llama-index-llms-gemini==0.4.10",
    "llama-index-llms-openai==0.3.18",
    "llama-index-multi-modal-llms-openai==0.4.3",
    "llama-index-program-openai==0.3.1",
    "llama-index-question-gen-openai==0.3.0",
    "llama-index-readers-file==0.4.4",
    "llama-index-readers-llama-parse==0.4.0",
    "llama-index-tools-google==0.3.0",
    "llama-parse==0.6.0",
    "markdown-it-py==3.0.0",
    "MarkupSafe==2.1.5",
    "marshmallow==3.26.1",
    "mdurl==0.1.2",
    "msal==1.31.1",
    "msal-extensions==1.2.0",
    "multidict==6.1.0",
    "mypy-extensions==1.0.0",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "nltk==3.9.1",
    "numpy==2.2.2",
    "oauthlib==3.2.2",
    "openai==1.61.1",
    "orjson==3.10.15",
    "packaging==24.2",
    "pandas==2.2.3",
    "paramiko==3.5.1",
    "pillow==10.4.0",
    "portalocker==2.10.1",
    "propcache==0.2.1",
    "proto-plus==1.26.0",
    "protobuf==5.29.3",
    "psutil==6.1.1",
    "pyaml==23.12.0",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.1",
    "pycparser==2.22",
    "pydantic==2.11.0a1",
    "pydantic_core==2.28.0",
    "pydub==0.25.1",
    "Pygments==2.19.1",
    "PyJWT==2.10.1",
    "PyNaCl==1.5.0",
    "pyparsing==3.2.1",
    "pypdf==5.3.0",
    "pyperclip==1.9.0",
    "Pysher==1.0.8",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "pytz==2025.1",
    "PyYAML==6.0.2",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "retrying==1.3.4",
    "rich==13.9.4",
    "rpds-py==0.22.3",
    "rsa==4.9",
    "ruff==0.9.6",
    "safehttpx==0.1.6",
    "semantic-version==2.10.0",
    "semver==3.0.4",
    "sentry-sdk==2.20.0",
    "sgmllib3k==1.0.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "SQLAlchemy==2.0.38",
    "starlette==0.45.3",
    "striprtf==0.0.26",
    "tenacity==9.0.0",
    "tiktoken==0.8.0",
    "tomlkit==0.13.2",
    "tqdm==4.67.1",
    "typer==0.15.1",
    "typing-inspect==0.9.0",
    "typing_extensions==4.12.2",
    "tzdata==2025.1",
    "uritemplate==4.1.1",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "websocket-client==1.8.0",
    "websockets==14.2",
    "wrapt==1.17.2",
    "yarl==1.18.3",
    "zipp==3.21.0",
    "Pillow>=10.0.0",
    "yt-dlp>=2025.1.27",
    "pydub>=0.25.1",
]

[tool.uv]
editable = []
